% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{affinity.asc}
\alias{affinity.asc}
\title{Sample map with affinities}
\format{
A ASC file. Original CRS UTM zone 17N.
}
\description{
Map representing the affinities (i.e. the likelihood of movement) across
each pixel in an area.
}
\examples{
(f <- system.file("extdata", "affinity.asc", package = "ConScapeRtools"))
r <- terra::rast(f)
terra::plot(r)

}
\seealso{
\link{suitability.asc} \cr
}
