% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{betweenness_qweighted}
\alias{betweenness_qweighted}
\title{Wrapper for the \code{betweenness_qweighted} function of \code{ConScape}}
\usage{
betweenness_qweighted(h)
}
\arguments{
\item{h}{\verb{[GridRSP]} \cr The output from the \code{\link[ConScapeR:GridRSP]{ConScapeR::GridRSP()}} function.}
}
\value{
\code{matrix}
}
\description{
Computes the quality-weighted betweenness for a ConScape.GridRSP object.
See Van Moorter et al. (2023, Methods in Ecology and Evolution) for more details.
}
\keyword{internal}
