% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{conscape}
\alias{conscape}
\title{ConScape jl function}
\usage{
conscape(
  data_dir,
  out_dir,
  r_target,
  r_source,
  r_res,
  landmark,
  theta,
  exp_d,
  iter
)
}
\value{
A Julia function for running ConScape
}
\description{
This is an internal function for running ConScape
}
\author{
Bill Peterman
}
\keyword{internal}
