% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{expected_cost}
\alias{expected_cost}
\title{Wrapper for the \code{expected_cost} function of \code{ConScape}}
\usage{
expected_cost(h)
}
\arguments{
\item{h}{\verb{[GridRSP]} \cr The output from the \code{\link[ConScapeR:GridRSP]{ConScapeR::GridRSP()}} function.}
}
\description{
Computes the RSP expected cost between all sources and non-zero targets.
See Van Moorter et al. (2023, Methods in Ecology and Evolution) for more details.
}
\keyword{internal}
