% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{GridRSP}
\alias{GridRSP}
\title{Wrapper for the GridRSP function of `ConScape`}
\usage{
GridRSP(g, theta)
}
\arguments{
\item{g}{`[Grid]` \cr The output from the [ConScapeR::Grid()] function.}

\item{theta}{`[numeric]` \cr The randomness parameter theta. Lower `theta` values
represent a more random walk, and higher values a more least-cost path behavior.}
}
\value{
A `ConScape.GridRSP` object within Julia.
}
\description{
Creates a GridRSP from a ConScape.Grid and the randomness parameter. Theta closer to zero gives a more random movement,
as theta increases the RSP distribution will converge to the optimal or least-cost paths.
Note that the computation becomes numerically unstable as theta becomes really small or large.
See Van Moorter et al. (2023, Methods in Ecology and Evolution) for more details.
}
\keyword{internal}
