% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tiles.R
\name{make_tiles}
\alias{make_tiles}
\title{Break raster into small tiles}
\usage{
make_tiles(
  tile_d,
  tile_trim,
  asc_dir = NULL,
  r,
  clear_dir = FALSE,
  landmark = 10L
)
}
\arguments{
\item{tile_d}{Dimensions (in meters) of tiles}

\item{tile_trim}{The amount of border to be trimmed from tiles after running ConScape (meters)}

\item{asc_dir}{Directory where .asc files of tiles will be written to. If NULL (Default), then files will be written to the temporary directory of the R session}

\item{r}{Raster file as `SpatRaster` to be broken up into smaller tiles}

\item{clear_dir}{Should existing files in the `asc_dir` be overwritten? This function must have an empty `asc_dir` to proceed}

\item{landmark}{The landmark value used for 'coarse_graining' with ConScape (Default = 10L). Used to determine which landscape tiles have data to be processed with ConScape}
}
\value{
A named list containing the `SpatVector`tiles created, the numeric identifier of tiles with usable data for ConScape, and the path to the directory where .asc tiles were written
}
\description{
This function will break a large raster into tiles of specified dimensions.
}
\details{
The smaller the tiles created, the faster each can be processed. The width of the `tile_trim` parameter will depend upon the movement settings of your ConScape run. If there are obvious tiling edges and artifacts in your final surfaces, then `tile_trim` needs to be increased.
}
\examples{
examples/make_tiles_example.R
}
\seealso{
[tile_rast()] for subsequent making of tiles
}
\author{
Bill Peterman
}
