% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{mat2rast}
\alias{mat2rast}
\title{Convert a matrix to raster}
\usage{
mat2rast(mat, rast)
}
\arguments{
\item{mat}{`[matrix]` \cr Matrix to be converted.}

\item{rast}{`[SpatRaster]` \cr Template raster, usually one of those used in the
[ConScapeR::Grid()] function.}
}
\value{
`[SpatRaster]`
}
\description{
Convert a matrix to raster
}
\keyword{internal}
