% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{patches.shp}
\alias{patches.shp}
\title{Sample shapefile of target patches}
\format{
A SHP file. Original CRS: UTM zone 17N.
}
\description{
A shapefile delineating potential habitat patches.
}
\examples{
(f <- system.file("extdata", "patches.shp", package = "ConScapeRtools"))
p <- terra::vect(f)
terra::plot(p)

}
