% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suitability.asc}
\alias{suitability.asc}
\title{Sample map with suitability}
\format{
A ASC file. Original CRS: UTM zone 17N.
}
\description{
Map representing the habitat quality in each pixel in an area.
}
\examples{
(f <- system.file("extdata", "suitability.asc", package = "ConScapeRtools"))
r <- terra::rast(f)
terra::plot(r)

}
\seealso{
\link{affinity.asc} \cr
}
