% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{vec2mat}
\alias{vec2mat}
\title{Convert a node vector to matrix}
\usage{
vec2mat(vec, g)
}
\arguments{
\item{vec}{`[matrix]` \cr Matrix to be converted}

\item{g}{`[Grid]` \cr Template raster, usually one of those used in the
[ConScapeR::Grid()] function.}
}
\value{
`[matrix]`
}
\description{
Convert a node vector to matrix
}
\keyword{internal}
