% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{betweenness_kweighted}
\alias{betweenness_kweighted}
\title{Wrapper for the \code{betweenness_kweighted} function of \code{ConScape}}
\usage{
betweenness_kweighted(h, alpha)
}
\arguments{
\item{h}{\verb{[GridRSP]} \cr The output from the \code{\link[ConScapeR:GridRSP]{ConScapeR::GridRSP()}} function.}

\item{alpha}{\verb{[numeric]} \cr The distance scaling for the exponential distance
to proximity transformation.}
}
\value{
\code{matrix}
}
\description{
Computes the quality- and proximity-weighted betweenness for a ConScape.GridRSP object.
See Van Moorter et al. (2023, Methods in Ecology and Evolution) for more details.
}
\keyword{internal}
