% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{connected_habitat}
\alias{connected_habitat}
\title{Wrapper for the \code{connected_habitat} function of \code{ConScape}}
\usage{
connected_habitat(h, alpha)
}
\arguments{
\item{h}{\verb{[GridRSP]} \cr The output from the \code{\link[ConScapeR:GridRSP]{ConScapeR::GridRSP()}} function.}

\item{alpha}{\verb{[numeric]} \cr The distance scaling for the exponential distance to
proximity transformation.}
}
\value{
A \code{matrix} with loca (pixel) measures of habitat functionality.
}
\description{
Computes the habitat functionality for a ConScape.GridRSP object.
See Van Moorter et al. (2023, Methods in Ecology and Evolution) for more details.
}
\keyword{internal}
