% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{mat2rast}
\alias{mat2rast}
\title{Convert a matrix to raster}
\usage{
mat2rast(mat, rast)
}
\arguments{
\item{mat}{\verb{[matrix]} \cr Matrix to be converted.}

\item{rast}{\verb{[SpatRaster]} \cr Template raster, usually one of those used in the
\code{\link[ConScapeR:Grid]{ConScapeR::Grid()}} function.}
}
\value{
\verb{[SpatRaster]}
}
\description{
Convert a matrix to raster
}
\keyword{internal}
