% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_conscape.R
\name{mosaic_conscape}
\alias{mosaic_conscape}
\title{Mosaic ConScape tiles into single contiguous raster}
\usage{
mosaic_conscape(asc_dir, tile_trim)
}
\arguments{
\item{tile_trim}{Provide `tile_trim` value used when running [ConScapeRtools:make_tiles()]}

\item{out_dir}{Directory where [ConScapeRtools:run_conscape()] results were written}
}
\value{
`SpatRaster`
}
\description{
After running [ConScapeRtools:run_conscape()], this function will reassemble tiles into a single `SpatRaster`.
}
\examples{
examples/mosaic_conscape_example.R
}
\seealso{
[make_tiles()] and [run_conscape()] \cr
}
\author{
Bill Peterman
}
