% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_conscape.R
\name{mosaic_conscape}
\alias{mosaic_conscape}
\title{Mosaic ConScape tiles into single contiguous raster}
\usage{
mosaic_conscape(out_dir, tile_trim, mask = NULL, method = c("mosaic", "merge"))
}
\arguments{
\item{out_dir}{Directory where \code{\link[=run_conscape]{run_conscape()}} results were written}

\item{tile_trim}{Provide \code{tile_trim} value used when running \code{\link[=conscape_prep]{conscape_prep()}} or \code{\link[=make_tiles]{make_tiles()}}}

\item{mask}{Default = NULL. \code{SpatRaster} indicating regions of the landscape that could potentially have pheasants}

\item{method}{Either 'mosaic' (Default) or 'merge' to combine raster tiles together}
}
\value{
\code{SpatRaster}
}
\description{
After running \code{\link[=run_conscape]{run_conscape()}}, this function will reassemble tiles into a single \code{SpatRaster}.
}
\examples{
library(ConScapeRtools)

## Import data
s <- system.file("data/suitability.asc", package = "ConScapeRtools")
source <- terra::rast(s)

a <- system.file("data/affinity.asc", package = "ConScapeRtools")
resist <- terra::rast(a)

jl_home <- "C:/Users/peterman.73/AppData/Local/Programs/Julia-1.10.5/bin/"

td <- tile_design(r_mov = resist,
                  r_target = source,
                  max_d = 7000,
                  theta = 0.1,
                  jl_home = jl_home)

## Tile dimension
tile_d <- td$tile_d

# How much to trim tiles
tile_trim <- td$tile_trim

# Makes computation more efficient
landmark <- 2L # Must be an integer, not numeric

# Controls level of randomness of paths
theta <- td$theta

# Controls rate of decay with distance
exp_d <- td$exp_d

## Prepare data for analysis
prep <- conscape_prep(tile_d = tile_d,
                      tile_trim = tile_trim,
                      r_target = source,
                      r_mov = resist,
                      r_src = source,
                      clear_dir = T,
                      landmark = landmark)

## Run ConScape
cs_run <- run_conscape(out_dir = file.path(prep$asc_dir,"results"),
                       conscape_prep = prep,
                       theta = theta,
                       exp_d = exp_d,
                       jl_home = jl_home,
                       parallel = T,
                       workers = 5,
                       end_multisession = T)

## Put output tiles together
cs_btwn <- mosaic_conscape(out_dir = cs_run$outdir_btwn,
                           tile_trim = tile_trim,
                           method = 'mosaic')
cs_fcon <- mosaic_conscape(out_dir = cs_run$outdir_fcon,
                           tile_trim = tile_trim)
plot(c(cs_btwn, cs_fcon))
}
\seealso{
\code{\link[=make_tiles]{make_tiles()}} and \code{\link[=run_conscape]{run_conscape()}} \cr
}
\author{
Bill Peterman
}
