% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_conscape.R
\name{run_conscape}
\alias{run_conscape}
\title{Run ConScape}
\usage{
run_conscape(
  data_dir,
  out_dir,
  hab_target,
  hab_src,
  mov_prob,
  landmark = 10L,
  theta = 0.01,
  exp_d = 150,
  jl_home,
  parallel = FALSE,
  workers = availableCores()/2,
  end_multisession = FALSE
)
}
\arguments{
\item{out_dir}{Directory where final ConScape outputs will be written}

\item{hab_target}{Directory with habitat target tiles}

\item{hab_src}{Directory with habitat source tiles}

\item{mov_prob}{Directory with movement probability tiles}

\item{landmark}{The landmark value used for 'coarse_graining' with ConScape (Default = 10L). Used to determine which landscape tiles have data to be processed with ConScape}

\item{theta}{Parameter to control the amount of randomness in the paths. As theta approaches 0 movement is random, whereas theta approaching infinity is optimal movement. (Default = 0.01)}

\item{exp_d}{Numerator of the exponential decay function used with the distance transformation of the movement grid (Default = 150)}

\item{jl_home}{Path to the `bin` directory where Julia is installed}

\item{parallel}{Logical. If FALSE, processing will not be done in parallel.}

\item{workers}{If `parallel = TRUE`, provide the number of parallel workers to create. Default 0.5*Number of available cores}

\item{end_multisession}{Logical. Default == FALSE}

\item{clear_dir}{Should existing files in the `asc_dir` be overwritten? This function must have an empty `asc_dir` to proceed}
}
\value{
A named list with the directories where betweenness and functional connectivity outputs were written
}
\description{
This function runs ConScape (optionally in parallel) over all created landscape tiles
}
\examples{
examples/run_conscape_example.R
}
\author{
Bill Peterman
}
