% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_design.R
\name{tile_design}
\alias{tile_design}
\title{Function to determine suitable tiling and analysis parameters}
\usage{
tile_design(r_mov, r_target, max_d, theta = 0.1, jl_home)
}
\arguments{
\item{r_mov}{`SpatRaster` representing movement probabilities}

\item{r_target}{`SpatRaster` representing landscape target quality}

\item{max_d}{Presumed maximum movement distance (measured in meters) through ideal habitat}

\item{theta}{Parameter to control the amount of randomness in the paths. As theta approaches 0 movement is random, whereas theta approaching infinity is optimal movement. (Default = 0.1)}
}
\value{
A named list with the path to the directory where .asc tiles were written
}
\description{
Used to determine the tile size, trim width, exp_d decay value given properties of data inputs.
}
\details{
`r_mov` SpatRaster must be in a projected coordinate reference system.
Relies on functions borrowed from the `ConScapeR` package (https://github.com/ConScape/ConScapeR).
}
\examples{
examples/tile_design_example.R
}
\author{
Bill Peterman
}
