% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_design.R
\name{tile_design}
\alias{tile_design}
\title{Function to determine suitable tiling and analysis parameters}
\usage{
tile_design(r_mov, r_target, max_d, theta = 0.1, jl_home)
}
\arguments{
\item{r_mov}{\code{SpatRaster} representing movement probabilities}

\item{r_target}{\code{SpatRaster} representing landscape target quality}

\item{max_d}{Presumed maximum movement distance (measured in meters) through ideal habitat}

\item{theta}{Parameter to control the amount of randomness in the paths. As theta approaches 0 movement is random, whereas theta approaching infinity is optimal movement. (Default = 0.1)}

\item{jl_home}{Path to the bin directory where Julia resides}
}
\value{
A named list with parameters for creating tiles with \code{\link[=conscape_prep]{conscape_prep()}} or \code{\link[=tile_rast]{tile_rast()}}
}
\description{
Used to determine the tile size, trim width, exp_d decay value given properties of data inputs.
}
\details{
\code{r_mov} SpatRaster must be in a projected coordinate reference system.
Relies on functions borrowed from the \code{ConScapeR} package (https://github.com/ConScape/ConScapeR).
}
\examples{
library(ConScapeRtools)

## Import data
s <- system.file("data/suitability.asc", package = "ConScapeRtools")
source <- terra::rast(s)

a <- system.file("data/affinity.asc", package = "ConScapeRtools")
resist <- terra::rast(a)

jl_home <- "C:/Users/peterman.73/AppData/Local/Programs/Julia-1.10.5/bin/"

td <- tile_design(r_mov = resist,
                  r_target = source,
                  max_d = 7000,
                  theta = 0.1,
                  jl_home = jl_home)
}
\author{
Bill Peterman
}
