% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_rast.R
\name{tile_rast}
\alias{tile_rast}
\title{Break raster into small tiles}
\usage{
tile_rast(r, make_tiles, out_dir, clear_dir = FALSE, progress = FALSE)
}
\arguments{
\item{r}{Raster file as \code{SpatRaster} to be broken up into smaller tiles}

\item{make_tiles}{Object created from  \code{\link[=make_tiles]{make_tiles()}} function}

\item{out_dir}{Directory where .asc files of tiles will be written. This should be different directory from the \code{make_tiles} \code{asc_dir}}

\item{clear_dir}{Should existing files in the \code{out_dir} be overwritten? This function must have an empty \code{out_dir} to proceed}

\item{progress}{Logical. If \code{TRUE}, processing progress will be reported.}
}
\value{
A named list with the path to the directory where .asc tiles were written
}
\description{
This uses objects created by the \code{\link[=make_tiles]{make_tiles()}} function to create identical tiles from other \code{SpatRaster} objects.
}
\examples{
library(ConScapeRtools)

## Import data
s <- system.file("data/suitability.asc", package = "ConScapeRtools")
source <- terra::rast(s)

a <- system.file("data/affinity.asc", package = "ConScapeRtools")
resist <- terra::rast(a)

jl_home <- "C:/Users/peterman.73/AppData/Local/Programs/Julia-1.10.5/bin/"

## Run tile design to get parameters
td <- tile_design(r_mov = resist,
                  r_target = source,
                  max_d = 7000,
                  theta = 0.1,
                  jl_home = jl_home)

target_tile <- make_tiles(tile_d = td$tile_d,
                          tile_trim = td$tile_trim,
                          r = source,
                          clear_dir = T)

mov_tile <- tile_rast(r = resist,
                      make_tiles = target_tile ,
                      out_dir = file.path(target_tile$asc_dir, 'mov'),
                      clear_dir = T)
}
\seealso{
\code{\link[=make_tiles]{make_tiles()}} for initial creation of tiles
}
\author{
Bill Peterman
}
