% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_rast.R
\name{tile_rast}
\alias{tile_rast}
\title{Break raster into small tiles}
\usage{
tile_rast(r, make_tiles, out_dir, clear_dir = FALSE)
}
\arguments{
\item{r}{Raster file as `SpatRaster` to be broken up into smaller tiles}

\item{make_tiles}{Object created from  [make_tiles()] function}

\item{out_dir}{Directory where .asc files of tiles will be written. This should be different directory from the `make_tiles` `asc_dir`}

\item{clear_dir}{Should existing files in the `out_dir` be overwritten? This function must have an empty `out_dir` to proceed}
}
\value{
A named list with the path to the directory where .asc tiles were written
}
\description{
This uses objects created by the [ConScapeRtools:make_tiles()] function to create identical tiles from other `SpatRaster` objects.
}
\examples{
examples/tile_rast_example.R
}
\seealso{
[make_tile()] for initial creation of tiles
}
\author{
Bill Peterman
}
