% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlConnectoR.R
\name{vec2mat}
\alias{vec2mat}
\title{Convert a node vector to matrix}
\usage{
vec2mat(vec, g)
}
\arguments{
\item{vec}{\verb{[matrix]} \cr Matrix to be converted}

\item{g}{\verb{[Grid]} \cr Template raster, usually one of those used in the
\code{\link[ConScapeR:Grid]{ConScapeR::Grid()}} function.}
}
\value{
\verb{[matrix]}
}
\description{
Convert a node vector to matrix
}
\keyword{internal}
